+++ indico/migrations/versions/20230424_1511_cd7038a37646_add_notes_column_to_room_bookings.py
@@ -1,0 +1,25 @@
+"""Add internal_note column to room bookings
+
+Revision ID: cd7038a37646
+Revises: 5d05eda06776
+Create Date: 2023-04-24 15:11:55.407209
+"""
+
+import sqlalchemy as sa
+from alembic import op
+
+
+# revision identifiers, used by Alembic.
+revision = 'cd7038a37646'
+down_revision = '5d05eda06776'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    op.add_column('reservations', sa.Column('internal_note', sa.Text(), nullable=False, server_default=''), schema='roombooking')
+    op.alter_column('reservations', 'internal_note', server_default=None, schema='roombooking')
+
+
+def downgrade():
+    op.drop_column('reservations', 'internal_note', schema='roombooking')