+++ indico/modules/rb/client/js/common/config/selectors.js
@@ -19,3 +19,4 @@ export const getTOSHTML = ({config}) => config.data.tosHTML;
 export const hasPrivacyPolicy = ({config}) => config.data.hasPrivacyPolicy;
 export const getPrivacyPolicyHTML = ({config}) => config.data.privacyPolicyHTML;
 export const getContactEmail = ({config}) => config.data.contactEmail;
+export const hasInternalNotesEnabled = ({config}) => config.data.internalNotesEnabled;